refactor: Standardize WithdrawalStatusModal components
