+++ src/components/Modals/WithdrawalStatusModal/CompleteWithdrawal/index.ts
@@ -1,0 +1,2 @@
+import CompleteWithdrawal from './CompleteWithdrawal'
+export default CompleteWithdrawal
