+++ src/components/Modals/WithdrawalStatusModal/LinkWrapper/LinkWrapper.types.ts
@@ -1,0 +1,4 @@
+export type Props = {
+  href: string
+  children: React.ReactNode
+}
