+++ src/components/Modals/WithdrawalStatusModal/LinkWrapper/index.ts
@@ -1,0 +1,2 @@
+import LinkWrapper from './LinkWrapper'
+export default LinkWrapper
