+++ src/components/Modals/WithdrawalStatusModal/ReadyToWithdraw/ReadyToWithdraw.types.ts
@@ -1,0 +1,5 @@
+import { Withdrawal } from "../../../../modules/mana/types";
+
+export type Props = {
+  withdrawal: Withdrawal
+}
