+++ src/components/Modals/WithdrawalStatusModal/ReadyToWithdraw/index.ts
@@ -1,0 +1,2 @@
+import ReadyToWithdraw from './ReadyToWithdraw'
+export default ReadyToWithdraw
