+++ src/components/Modals/WithdrawalStatusModal/WithdrawInitialized/WithdrawInitialized.types.ts
@@ -1,0 +1,5 @@
+import { Withdrawal } from "../../../../modules/mana/types";
+
+export type Props = {
+  withdrawal: Withdrawal
+}
