+++ src/components/Modals/WithdrawalStatusModal/WithdrawInitialized/index.ts
@@ -1,0 +1,2 @@
+import WithdrawInitialized from './WithdrawInitialized'
+export default WithdrawInitialized
